my_requirements.txt
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-doc==0.0.4
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==2.1.1
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
crcmod==1.7
cryptography==46.0.5
cycler==0.12.1
defusedxml==0.7.1
einops==0.6.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
filelock==3.14.0
fonttools==4.61.1
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
intel-openmp==2021.4.0
Jinja2==3.1.6
jmespath==0.10.0
json5==0.13.0
kiwisolver==1.4.9
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mkl==2021.4.0
mmcv-full==1.7.0
mmengine==0.10.7
model-index==0.1.11
mpmath==1.3.0
networkx==3.4.2
ninja==1.13.0
numpy==1.26.4
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
opendatalab==0.0.10
openmim==0.3.9
openpyxl==3.1.5
openxlab==0.1.3
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.3.3
Pillow==9.5.0
platformdirs==4.9.2
pycocotools==2.0.11
pycparser==3.0
pycryptodome==3.23.0
pyDeprecate==0.4.0
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2023.4
pywin32==311
PyYAML==6.0.1
regex==2026.2.19
requests==2.28.2
rich==13.4.2
safetensors==0.7.0
scipy==1.15.3
seaborn==0.13.2
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
supervision==0.27.0.post1
sympy==1.14.0
tabulate==0.9.0
tbb==2021.13.1
termcolor==3.3.0
terminaltables==3.1.10
timm==0.9.12
tokenizers==0.13.3
tomli==2.4.0
torch==1.13.1+cpu
torchaudio==0.13.1+cpu
torchvision==0.14.1+cpu
tqdm==4.65.2
transformers==4.30.2
typer==0.24.1
typer-slim==0.24.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==1.26.13
yapf==0.43.0
